More responsiveness.

+++ src/main/java/com/Lukas/demo/controller/apiController.java
@@ -6,6 +6,7 @@ import org.hibernate.validator.constraints.Length;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
@@ -25,6 +26,9 @@ public class apiController
     @Autowired
     private ToDoItemRepository toDoItemRepository;
 
+    @Autowired
+    private SimpMessagingTemplate template;
+
     @GetMapping("/hello")
     public String hello(@RequestParam(value = "name", defaultValue = "World") String name) {
         return String.format("Hello %s!", name);
@@ -41,16 +45,18 @@ public class apiController
         n.setName(name);
         n.setDone(done);
         toDoItemRepository.save(n);
+        template.convertAndSend("/todo/sub/add", n);
         return "Saved";
     }
 
     @PostMapping("/update")
-    public ResponseEntity updateToDo(@RequestParam @Min(0) Long id, @RequestParam @Max(20) String name, @RequestParam Boolean done) {
+    public ResponseEntity updateToDo(@RequestParam @Min(0) Long id, @RequestParam @Size(min = 2, max = 30) String name, @RequestParam Boolean done) {
         ToDoItem n = toDoItemRepository.findById(id).orElse(null);
         if(n != null) { //backup validation, @min and @max should cover it though.
             n.setName(name);
             n.setDone(done);
             toDoItemRepository.save(n);
+            template.convertAndSend("/todo/sub/update", n);
             return new ResponseEntity(HttpStatus.OK);
         } else {
             return new ResponseEntity(HttpStatus.NOT_FOUND);
@@ -59,8 +65,13 @@ public class apiController
 
     @PostMapping("/delete")
     public ResponseEntity deleteToDo(@RequestParam @Min(0) Long id) {
-        toDoItemRepository.deleteById(id);
-        return new ResponseEntity(HttpStatus.OK);
+        if(toDoItemRepository.existsById(id)) {
+            ToDoItem i = toDoItemRepository.findById(id).get(); // GetById causes json convert error
+            toDoItemRepository.deleteById(id);
+            template.convertAndSend("/todo/sub/delete", i);
+            return new ResponseEntity(HttpStatus.OK);
+        }
+        return new ResponseEntity(HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping("/template")
